Add: 自分のメッセージ以外は削除できない機能を追加。
messageのuser_idがcurrent_user.idかどうかを判別する式を追加しました。

+++ app/controllers/users/messages_controller.rb
@@ -19,8 +19,11 @@ class Users::MessagesController < ApplicationController
 
   def destroy
     message = Message.find(params[:id])
-    message.destroy
-    redirect_to users_room_path(message.room_id)
+    # 自分のメッセージ以外は削除できないようにする
+    if message.user_id == current_user.id
+      message.destroy
+      redirect_to users_room_path(message.room_id)
+    end
   end
 
   private
